refactor(PhoneBox): autoHyphen함수 변경, enter클릭 시 유효성 검사 후 다음 탭으로 이동
