+++ src/components/signup/PhoneBox.tsx
@@ -1,14 +1,31 @@
 import { useFormContext } from 'react-hook-form';
 import { UserSignupType } from '@/types/country/user';
-import { autoHyphen } from '@/utils/phone';
+import { autoHyphen } from '@/utils/client/phone';
 import { error, inputContent } from '@/styles/signup/page.css';
 import FormField from './FormField';
 
 const PhoneBox = () => {
   const {
     register,
+    setFocus,
+    trigger,
     formState: { errors },
   } = useFormContext<UserSignupType>();
+
+  const handleKeyDown = async (
+    event: React.KeyboardEvent<HTMLInputElement>,
+  ) => {
+    if (event.key === 'Enter') {
+      if (event.nativeEvent.isComposing) {
+        return;
+      }
+      event.preventDefault();
+      const isValid = await trigger('phoneNumber');
+      if (isValid) {
+        setFocus('selfIntroduction');
+      }
+    }
+  };
 
   return (
     <FormField title="전화번호" required={true}>
@@ -17,18 +34,20 @@ const PhoneBox = () => {
           state: !!errors.phoneNumber ? 'error' : 'default',
         })}
         {...register('phoneNumber', {
+          onChange: (e) => {
+            e.target.value = autoHyphen(e.target.value);
+          },
           required: '전화번호를 입력해주세요.',
           pattern: {
-            value: /^01([0|1|6|7|8|9])-?([0-9]{3,4})-?([0-9]{4})$/,
+            value: /^\d{3}-\d{4}-\d{4}$/,
             message: '올바른 전화번호 형식이 아닙니다.',
           },
         })}
-        onChange={(e) => {
-          e.target.value = autoHyphen(e.target.value);
-        }}
         type="text"
+        tabIndex={8}
         maxLength={13}
         placeholder="전화번호를 입력하세요."
+        onKeyDown={handleKeyDown}
       />
       {errors?.phoneNumber && (
         <div className={error}>{errors.phoneNumber.message}</div>
