+++ src/utils/client/phone.ts
@@ -1,18 +1,10 @@
 export const autoHyphen = (value: string) => {
   const numbers = value.replace(/[^0-9]/g, '');
-
-  if (numbers.length < 4) {
+  if (numbers.length <= 3) {
     return numbers;
+  } else if (numbers.length <= 7) {
+    return `${numbers.slice(0, 3)}-${numbers.slice(3)}`;
+  } else {
+    return `${numbers.slice(0, 3)}-${numbers.slice(3, 7)}-${numbers.slice(7)}`;
   }
-  if (numbers.length < 7) {
-    return numbers.slice(0, 3) + '-' + numbers.slice(3);
-  }
-  if (numbers.length < 11) {
-    return (
-      numbers.slice(0, 3) + '-' + numbers.slice(3, 6) + '-' + numbers.slice(6)
-    );
-  }
-  return (
-    numbers.slice(0, 3) + '-' + numbers.slice(3, 7) + '-' + numbers.slice(7, 11)
-  );
 };
